+++ demo/api/src/app.module.ts
@@ -105,6 +105,9 @@ import { RedirectScope } from "./redirects/dto/redirect-scope";
             useFactory: async (config: ConfigType<typeof configNS>) => ({
                 config: {
                     helmRelease: config.HELM_RELEASE,
+                    isContentScopeAllowed: () => {
+                        return true;
+                    },
                 },
             }),
             inject: [configNS.KEY],