+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## NEXT
+## 3.2.0
+
+_Nov 29, 2022_
+
+### Highlights
+
+-   Add support to type root blocks in the GraphQL schema using `RootBlockDataScalar` and `RootBlockInputScalar`. See [page.entity.ts](/demo/api/src/pages/entities/page.entity.ts), [schema.gql](/demo/api/schema.gql), and [codegen.ts](demo/admin/codegen.ts) for an example on how to use
 
 ### @comet/admin-rte
 
@@ -12,12 +18,16 @@ All notable changes to this project will be documented in this file. This projec
 
 ### @comet/cms-admin
 
+#### Changes
+
 -   Fix a bug where the pages query would query for a field `undefined` when no `additionalPageTreeNodeFragment` is set
 
 ### @comet/cms-api
 
+#### Changes
+
 -   Add default value `{}` for `RedirectScopeInput` when no explicit scope is set to make redirects scope support backwards compatible
--   Add `RootBlockDataScalar` and `RootBlockInputScalar` to type root blocks in the GraphQL schema. See [page.entity.ts](/demo/api/src/pages/entities/page.entity.ts), [schema.gql](/demo/api/schema.gql), and [codegen.ts](demo/admin/codegen.ts) for an example on how to use
+-   Add `RootBlockDataScalar` and `RootBlockInputScalar` scalars to type root blocks in the GraphQL schema
 
 ## 3.1.0